refactoring 1: put code into package
